+++ community/kernel/src/main/java/org/neo4j/kernel/api/properties/ByteProperty.java
@@ -48,7 +48,7 @@ final class ByteProperty extends NumberPropertyWithin4Bytes
     {
         if ( other instanceof Byte )
         {
-            return other == value;
+            return other == (byte)value;
         }
 
         return valueCompare( value, other );
