+++ community/kernel/src/main/java/org/neo4j/kernel/api/properties/SmallLongProperty.java
@@ -48,7 +48,7 @@ final class SmallLongProperty extends NumberPropertyWithin4Bytes
     {
         if ( other instanceof Integer )
         {
-            return value == other;
+            return value == (int)other;
         }
 
         return valueCompare( value, other );
